Pedidos ahora se insertan correctamente con los clientes_id asociados al email logueado en la aplicacion

+++ app/src/main/java/com/delivery_clientes/data/repository/DetalleRepository.java
@@ -23,7 +23,7 @@ public class DetalleRepository {
 
     public void insertarDetallePedido(PedidoDetalle ped) {pedidosDetalleDAO.insertarPedidoDetalle(ped);}
 
-    public LiveData<List<PedidoDetalle>> obtenerDetalleLive(int id){
+    public LiveData<List<PedidoDetalle>> obtenerDetalleLive(long id){
         return pedidosDetalleDAO.obtenerDetallePedidoPorIdPedido(id);
     }
